Integrated everyons work and cleaned up file structure
Things were getting messy with a few files being outside both client and server. I moved them to their correct spot and everything should be working. I also added everyones work to app.js in client so it's all visible and accessible. some ports were set to 5001 for the backend I did change those to 5000 just to get it working on my end but it should be easy to change it back on anyone system if that port is causing problems.